Class_07_LLM/pyproject.toml
[project]
name = "class-07-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Unicorn4rich", email = "sk9952908@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "ruff>=0.9.9",
    "streamlit>=1.42.2",
]

[project.scripts]
class-07-llm = "class_07_llm:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
